Move 2 test cases from chatbots utils to leetcode utils
